requirements.txt
Flask==2.3.3
bcrypt==4.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
Werkzeug==2.3.7
Jinja2==3.1.2
